bluesky-assign3/requirements.txt
scikit-learn
pandas
numpy
atproto
python-dotenv
openai
spacy
